+++ tests/testthat/test-one_two_sample_dataframes.R
@@ -1,10 +1,11 @@
-
 test_that(
-  desc = " dataframes for parametric t-tests",
+  desc = " parametric t-tests",
   code = {
     skip_if(getRversion() < "4.0")
+    options(tibble.width = Inf)
 
-    # dataframes for one-sample t-test (with NAs) ---------
+    # one-sample t-test (with NAs) ---------
+
     set.seed(123)
     df_1 <-
       purrr::pmap_dfr(
@@ -13,6 +14,7 @@ test_that(
           x = list("brainwt"),
           test.value = list(0.25),
           effsize.type = list("d", "g", "d", "g"),
+          alternative = c("two.sided", "less", "greater", "two.sided"),
           var.equal = list(TRUE, FALSE, TRUE, FALSE),
           conf.level = list(0.89, 0.99, 0.90, 0.50)
         ),
@@ -22,7 +24,8 @@ test_that(
     set.seed(123)
     expect_snapshot(df_1)
 
-    # dataframes for parametric t-test (between-subjects without NAs) ---------
+    # parametric t-test (between-subjects without NAs) ---------
+
     set.seed(123)
     df_2_between <-
       purrr::pmap_dfr(
@@ -32,6 +35,7 @@ test_that(
           y = list("wt"),
           effsize.type = list("d", "g", "d", "g"),
           var.equal = list(TRUE, FALSE, TRUE, FALSE),
+          alternative = c("two.sided", "less", "greater", "two.sided"),
           conf.level = list(0.89, 0.99, 0.90, 0.50)
         ),
         .f = statsExpressions::two_sample_test
@@ -40,7 +44,8 @@ test_that(
     set.seed(123)
     expect_snapshot(df_2_between)
 
-    # dataframes for parametric t-test (within-subjects with NAs) ---------
+    # parametric t-test (within-subjects with NAs) ---------
+
     set.seed(123)
     df_2_within <-
       purrr::pmap_dfr(
@@ -58,5 +63,43 @@ test_that(
 
     set.seed(123)
     expect_snapshot(df_2_within)
+
+    # parametric ANOVA (within-subjects with NAs) ---------
+
+    set.seed(123)
+    df_3_between <-
+      purrr::pmap_dfr(
+        .l = list(
+          data = list(ggplot2::msleep),
+          x = list("vore"),
+          y = list("sleep_rem"),
+          effsize.type = list("eta", "omega", "eta", "omega"),
+          var.equal = list(TRUE, FALSE, TRUE, FALSE),
+          conf.level = list(0.89, 0.80, 0.90, 0.50)
+        ),
+        .f = statsExpressions::oneway_anova
+      )
+
+    set.seed(123)
+    expect_snapshot(df_3_between)
+
+    # parametric ANOVA (within-subjects with NAs) ---------
+
+    set.seed(123)
+    df_3_within <-
+      purrr::pmap_dfr(
+        .l = list(
+          data = list(bugs_long),
+          x = list("condition"),
+          y = list("desire"),
+          paired = list(TRUE),
+          effsize.type = list("eta", "omega"),
+          conf.level = list(0.89, 0.90)
+        ),
+        .f = statsExpressions::oneway_anova
+      )
+
+    set.seed(123)
+    expect_snapshot(df_3_within)
   }
 )