+++ tests/testthat/test-two_sample_parametric.R
@@ -1,21 +1,18 @@
-# parametric t-test (between-subjects without NAs) ---------------------------
-
 test_that(
   desc = "parametric t-test works (between-subjects without NAs)",
   code = {
+    # parametric t-test (between-subjects without NAs) ----------------------
 
+    options(tibble.width = Inf)
 
     # `statsExpressions` output
     set.seed(123)
-    using_function1 <-
+    df1 <-
       suppressWarnings(
         two_sample_test(
-          data = dplyr::filter(
-            movies_long,
-            genre == "Action" | genre == "Drama"
-          ),
-          x = genre,
-          y = rating,
+          ToothGrowth,
+          x = supp,
+          y = len,
           effsize.type = "d",
           var.equal = TRUE,
           conf.level = 0.99,
@@ -23,60 +20,26 @@ test_that(
         )
       )
 
-    # expected output
+    # testing all details
     set.seed(123)
-    results1 <-
-      ggplot2::expr(
-        paste(
-          italic("t")["Student"],
-          "(",
-          "612",
-          ") = ",
-          "-10.52948",
-          ", ",
-          italic("p"),
-          " = ",
-          "6.0984e-24",
-          ", ",
-          widehat(italic("d"))["Cohen"],
-          " = ",
-          "-0.92473",
-          ", CI"["99%"],
-          " [",
-          "-1.16064",
-          ", ",
-          "-0.68822",
-          "]",
-          ", ",
-          italic("n")["obs"],
-          " = ",
-          "614"
-        )
-      )
-
-    # testing overall call
-    expect_equal(using_function1$expression[[1]], results1)
+    expect_snapshot(dplyr::select(df1, -expression))
+    expect_snapshot(df1$expression[[1]])
   }
 )
-
-# parametric t-test (between-subjects with NAs) ------------------------------
 
 test_that(
   desc = "parametric t-test works (between-subjects with NAs)",
   code = {
-
+    # parametric t-test (between-subjects with NAs) --------------------------
 
     # `statsExpressions` output
     set.seed(123)
-    using_function1 <-
+    df1 <-
       suppressWarnings(
         two_sample_test(
-          data = dplyr::filter(
-            movies_long,
-            genre == "Action" | genre == "Drama"
-          ),
-          x = genre,
-          y = rating,
+          ToothGrowth,
+          x = supp,
+          y = len,
           effsize.type = "g",
           var.equal = FALSE,
           conf.level = 0.90,
@@ -84,52 +47,21 @@ test_that(
         )
       )
 
-    # expected output
+    # testing all details
     set.seed(123)
-    results1 <-
-      ggplot2::expr(
-        paste(
-          italic("t")["Welch"],
-          "(",
-          "271.302",
-          ") = ",
-          "-9.275",
-          ", ",
-          italic("p"),
-          " = ",
-          "5.8e-18",
-          ", ",
-          widehat(italic("g"))["Hedges"],
-          " = ",
-          "-0.924",
-          ", CI"["90%"],
-          " [",
-          "-1.074",
-          ", ",
-          "-0.773",
-          "]",
-          ", ",
-          italic("n")["obs"],
-          " = ",
-          "614"
-        )
-      )
-
-    # testing overall call
-    expect_equal(using_function1$expression[[1]], results1)
+    expect_snapshot(dplyr::select(df1, -expression))
+    expect_snapshot(df1$expression[[1]])
   }
 )
-
-# parametric t-test (within-subjects without NAs) ---------------------------
 
 test_that(
   desc = "parametric t-test works (within-subjects without NAs)",
   code = {
-
+    # parametric t-test (within-subjects without NAs) -----------------------
 
     # output from `statsExpressions` helper subtitle
     set.seed(123)
-    subtitle <-
+    df1 <-
       suppressWarnings(two_sample_test(
         data = dplyr::filter(
           iris_long,
@@ -143,52 +75,22 @@ test_that(
         conf.level = 0.50
       ))
 
-    # expected
-    expected <-
-      ggplot2::expr(
-        paste(
-          italic("t")["Student"],
-          "(",
-          "149",
-          ") = ",
-          "34.8152",
-          ", ",
-          italic("p"),
-          " = ",
-          "1.85e-73",
-          ", ",
-          widehat(italic("g"))["Hedges"],
-          " = ",
-          "2.8283",
-          ", CI"["50%"],
-          " [",
-          "2.7086",
-          ", ",
-          "2.9560",
-          "]",
-          ", ",
-          italic("n")["pairs"],
-          " = ",
-          "150"
-        )
-      )
-
-    # testing overall call
-    expect_identical(subtitle$expression[[1]], expected)
+    # testing all details
+    set.seed(123)
+    expect_snapshot(dplyr::select(df1, -expression))
+    expect_snapshot(df1$expression[[1]])
   }
 )
 
 
-# parametric t-test (within-subjects with NAs) ---------------------------
-
 test_that(
   desc = "parametric t-test works (within-subjects with NAs)",
   code = {
-
+    # parametric t-test (within-subjects with NAs) ---------------------------
 
     # output from `statsExpressions` helper subtitle
     set.seed(123)
-    subtitle <-
+    df1 <-
       two_sample_test(
         data = dplyr::filter(bugs_long, condition %in% c("HDHF", "HDLF")),
         x = condition,
@@ -198,47 +100,17 @@ test_that(
         k = 3
       )
 
-    # expected
-    expected <-
-      ggplot2::expr(
-        paste(
-          italic("t")["Student"],
-          "(",
-          "89",
-          ") = ",
-          "3.613",
-          ", ",
-          italic("p"),
-          " = ",
-          "5e-04",
-          ", ",
-          widehat(italic("d"))["Cohen"],
-          " = ",
-          "0.381",
-          ", CI"["95%"],
-          " [",
-          "0.167",
-          ", ",
-          "0.597",
-          "]",
-          ", ",
-          italic("n")["pairs"],
-          " = ",
-          "90"
-        )
-      )
-
-    # testing overall call
-    expect_identical(subtitle$expression[[1]], expected)
+    # testing all details
+    set.seed(123)
+    expect_snapshot(dplyr::select(df1, -expression))
+    expect_snapshot(df1$expression[[1]])
   }
 )
-
-# works with subject id ------------------------------------------------------
 
 test_that(
   desc = "works with subject id",
   code = {
-
+    # works with subject id ------------------------------------------------
 
     # data
     df <-